backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
supabase==2.3.4
python-dotenv==1.0.0
pyjwt==2.8.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
